moved/renamed/created Common.Gw2 ChatCode
